+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Melee weapons without a long range, imported with it set to 5, now empty.
 * Crossbow bolts had range set to 5/5
 * Magic weapons had stopped applying magic bonus to damage
-
+* Duergar Mind Master did not parse special actions properly
 
 # 0.6.15